Electromagnetism and Image Folder
Added notes on electromagnetism and the hand rules. Also added a folder where images can be stored to be used in the notes. (I was taking screenshots of calculations since it's a lot faster and more concise in the notes since it goes from left to right instead of up and down like mathjax)